made services throw exceptions

+++ src/main/java/com/transactionsexample/spring_transactions/service/TransactionService.java
@@ -1,6 +1,9 @@
 package com.transactionsexample.spring_transactions.service;
 
+import com.transactionsexample.spring_transactions.Utils;
 import com.transactionsexample.spring_transactions.dto.TransactionDTO;
+import com.transactionsexample.spring_transactions.exceptions.InvalidTransactionException;
+import com.transactionsexample.spring_transactions.exceptions.TransactionNotFoundException;
 import com.transactionsexample.spring_transactions.repository.TransactionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,13 +20,19 @@ public class TransactionService {
         this.transactionRepository = transactionRepository;
     }
 
-    public TransactionDTO createTransaction(TransactionDTO transactionDTO) {
+    public void createTransaction(TransactionDTO transactionDTO) {
+        if(transactionDTO.getAmount() <= 0){
+            throw new InvalidTransactionException("Transaction must have an amount bigger than 0");
+        }
+        transactionDTO.setId(Utils.generateRandomId());
         transactionRepository.save(transactionDTO);
-        return transactionDTO;
     }
 
     public TransactionDTO getTransactionById(Long id) {
-        return transactionRepository.findById(id);
+        TransactionDTO transaction = transactionRepository.findById(id);
+        if(transaction == null)
+            throw new TransactionNotFoundException("Transaction not found");
+        return transaction;
     }
 
     public List<TransactionDTO> getAllTransactions() {
@@ -31,6 +40,9 @@ public class TransactionService {
     }
 
     public void deleteTransactionById(Long id) {
+        if(!transactionRepository.existsById(id)){
+            throw new TransactionNotFoundException("No transaction associated with id: "+ id);
+        }
         transactionRepository.deleteById(id);
     }
